fix component
