+++ initial-application/src/assets/PocketLogo.png
