+++ initial-application/src/components/LogoAndButton/index.css
@@ -1,0 +1,19 @@
+.app-container {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.logo {
+  height: 15rem;
+  margin-top: 2rem;
+}
+
+a,
+a:link,
+a:visited,
+a:active {
+  text-decoration: none;
+}
